Add one final day 22 solution plus some notes
This is a slightly smarter version of the naive approach with a (bit)
array.  Works fine on the AOC input, but it and doit2a.cc struggle on
bigger examples.  The original explicit cube cutting in doit2.cc and
the BDD approach in doit2b.cc both seem to scale better.